Homework 6: fixs crash, launch service and track location

+++ app/src/main/java/ru/fefu/activitytracker/db/entity/Activities.kt
@@ -15,7 +15,7 @@ data class Activities(
 @Entity
 data class Coordinates(
     @PrimaryKey(autoGenerate = true) val coordinate_id: Int,
-    @ColumnInfo(name = "activity_id") val activityId: Int,
+    @ColumnInfo(name = "activity_id") val activityId: Long,
     val longitude: Double,
     val latitude: Double,
 )
